requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.12
python-dotenv==1.2.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.1
dnspython>=2.0.0

